third_party/blink/web_tests/platform/generic/http/tests/eventsource/eventsource-bad-mime-type-expected.txt

CONSOLE ERROR: EventSource's response has a MIME type ("text/bogus") that is not "text/event-stream". Aborting the connection.
Test EventSource with an event-stream with incorrect mime-type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS es.readyState is es.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
